Packaged into a src directory
Skeleton was puzzled that even with a non-editable install the test
workflow was passing while the wheel workflow was failing. This article
cleared up the confusion: https://hynek.me/articles/testing-packaging/

It turns out that the current working directory gets searched first
before installed packages, so pytest was always testing the source
directory, not the installed directory. This was only picked up when the
wheel was installed in an isolated environment in Github Actions.

Moving to a src directory allowed the problem to be reproduced locally:
  pipenv run pip install .
  pipenv run tests
  pipenv run skeleton-says Hello