Added mypy
Skeleton doesn't have a great memory, and was worried that Python's
dynamic types would be confusing long term. Adding type hints answers
the question "what type of object does this function return" in a
statically checkable way that is IDE friendly. It can also be added
gradually to a codebase.

Run checks with:
  pipenv run mypy .