Release to PyPI on tag
Skeleton's customer wanted an easier deploy mechanism, and had heard
about PyPI. Skeleton made an account on PyPI, generated an API token,
then added it to the Secrets of the repository as PYPI_TOKEN as shown in
z_slide.png. This was then used in the workflow to make a release to
PyPI on tag.

Some orgs (like dls-controls) already have the PYPI_TOKEN set at org
level, so they do not need to be added to each repo.